refactor: add function to inject css from outside
