--- packages/core/src/components/index.ts
@@ -1,4 +1,0 @@
-export * from './AutoLayout'
-export * from './Box'
-export * from './Flex'
-export * from './Stack'
