--- packages/core/src/contexts/MediaQuery/index.ts
@@ -1,1 +1,0 @@
-export type OptionWithMediaQuery<T extends { [property: string]: any }> = { [key in keyof T]: T[key][] | T[key] }
