+++ packages/core/src/index.ts
@@ -1,14 +1,13 @@
-export { error } from './common'
-export type { AsProps, AxisDirection, CSSPixelValue } from './common'
 export type {
-  AutoLayoutComponentType,
-  AutoLayoutProps,
-  BoxComponentType,
+  BoxOption,
   BoxProps,
-  FlexComponentType,
+  BoxComponentType,
+  FlexOption,
   FlexProps,
-  StackComponentType,
+  FlexComponentType,
+  StackOption,
   StackProps,
-} from './components'
-export type { AutoLayoutOption, BoxOption, FlexOption, GutterOption, StackOption, XAxis, YAxis } from './utils'
-export type { OptionWithMediaQuery } from './contexts'
+  StackComponentType,
+} from './layout-components'
+export { createUtils } from './utils'
+export { creatMediaQuerySet } from './responsive'
