--- packages/core/src/utils/flex/index.ts
@@ -1,7 +1,0 @@
-import { CSSProperties } from 'react'
-
-export interface FlexOption {
-  align?: CSSProperties['alignItems']
-  justify?: CSSProperties['justifyContent']
-  direction?: CSSProperties['flexDirection']
-}
