--- packages/core/src/utils/gutter/index.ts
@@ -1,5 +1,0 @@
-import { CSSProperties } from 'react'
-
-export interface GutterOption {
-  spacing?: CSSProperties['gap']
-}
