+++ packages/emotion/src/layout-components/Stack.tsx
@@ -1,9 +1,8 @@
 import { ComponentPropsWithRef, ElementType, forwardRef, useContext } from 'react'
-import { css } from '@emotion/react'
-import { AxisDirection, StackComponentType, StackOption, StackProps } from '@jsxcss/core'
+import { StackComponentType, StackOption, StackProps } from '@jsxcss/core'
 import { Box } from './Box'
 import { MediaQueryContext } from '../responsive'
-import { flex, gutter } from '../utils'
+import { stack } from '../utils'
 
 const createStack = (defaultOption: StackOption = {}): StackComponentType =>
   forwardRef(function Stack<T extends ElementType>(props: StackProps<T>, ref: ComponentPropsWithRef<T>['ref']) {
@@ -15,20 +14,10 @@ const createStack = (defaultOption: StackOption = {}): StackComponentType =>
       spacing = defaultOption.spacing,
       align = defaultOption.align,
       justify = defaultOption.justify,
-
       ...rest
     } = props
 
-    const flexDirection = Array.isArray(direction) ? getFlexDirectionInArray(direction) : getFlexDirection(direction)
-
-    return (
-      <Box
-        as={as}
-        ref={ref}
-        css={css(flex({ direction: flexDirection, align, justify }, mediaQuery), gutter({ spacing }, mediaQuery))}
-        {...rest}
-      />
-    )
+    return <Box as={as} ref={ref} css={stack({ direction, align, justify, spacing }, mediaQuery)} {...rest} />
   })
 
 type StackType = StackComponentType & {
@@ -39,18 +28,3 @@ type StackType = StackComponentType & {
 export const Stack = createStack({ direction: 'vertical' }) as StackType
 Stack.Vertical = createStack({ direction: 'vertical' })
 Stack.Horizontal = createStack({ direction: 'horizontal' })
-
-const getFlexDirectionInArray = (axisDirections: (AxisDirection | undefined)[]) => axisDirections.map(getFlexDirection)
-const getFlexDirection = (axisDirection?: AxisDirection) => {
-  switch (axisDirection) {
-    case 'horizontal': {
-      return 'row'
-    }
-    case 'vertical': {
-      return 'column'
-    }
-    default: {
-      return undefined
-    }
-  }
-}
