+++ packages/emotion/src/layout-components/index.ts
@@ -1,4 +1,3 @@
 export * from './Box'
 export * from './Flex'
 export * from './Stack'
-export * from './AutoLayout'
