+++ packages/emotion/src/utils/index.ts
@@ -1,3 +1,4 @@
-export * from './box'
-export * from './flex'
-export * from './gutter'
+import { css } from '@emotion/react'
+import { createUtils } from '@jsxcss/core'
+
+export const { box, flex, gutter, stack } = createUtils(css)
