--- websites/emotion/CHANGELOG.md
@@ -1,43 +1,0 @@
-# @jsxcss/test-emotion
-
-## null
-
-### Patch Changes
-
-- Updated dependencies [7b24db5]
-  - @jsxcss/emotion@1.3.3
-
-## null
-
-### Patch Changes
-
-- Updated dependencies [f3c7f91]
-  - @jsxcss/emotion@1.3.2
-
-## null
-
-### Patch Changes
-
-- Updated dependencies [233932a]
-  - @jsxcss/emotion@1.3.1
-
-## null
-
-### Patch Changes
-
-- Updated dependencies [29e6952]
-  - @jsxcss/emotion@1.3.0
-
-## null
-
-### Patch Changes
-
-- Updated dependencies [ad3864c]
-  - @jsxcss/emotion@1.2.0
-
-## null
-
-### Patch Changes
-
-- Updated dependencies [49f5a28]
-  - @jsxcss/emotion@1.1.0
